final-project p2: реализована логика, без добавления статистики

+++ ewm-main/src/main/java/ru/practicum/controller/adminApi/AdminCategoriesController.java
@@ -22,13 +22,13 @@ public class AdminCategoriesController {
     }
 
     @DeleteMapping("/{catId}")
-    public void deleteCategory(@PathVariable Integer catId) {
+    public void deleteCategory(@PathVariable Long catId) {
         log.info("DELETE /admin/categories/{} is accessed", catId);
         categoriesService.deleteCategory(catId);
     }
 
     @PatchMapping("/{catId}")
-    public CategoryDto updateCategory(@PathVariable Integer catId, @RequestBody CategoryDto categoryDto) {
+    public CategoryDto updateCategory(@PathVariable Long catId, @RequestBody CategoryDto categoryDto) {
         log.info("PATCH /admin/categories/{}  is accessed: {}", catId, categoryDto);
         return categoriesService.updateCategory(catId, categoryDto);
     }
